Algorithms:
This was a test that I completed. I had 45 minutes to figure out how to
obtain the largest subset string that did not contain digits and had to
contain a uppercase character. This example serves as reference for the
cpp language as well as parsing through a string for specific characters. The use of standard functions such as isupper() and is islower() improve efficiency and readability.